server/requirements.txt
flask==2.0.3
werkzeug==2.0.3
pymongo==4.3.3
python-dotenv==0.19.2
pyjwt==2.3.0
requests==2.26.0
urllib3==1.26.6
